fix: delete tooltips on component unmount
