+++ frontend/src/scenes/funnels/FunnelBarVertical/FunnelBarVertical.tsx
@@ -26,8 +26,7 @@ export function FunnelBarVertical({ showPersonsModal: showPersonsModalProp = tru
     const { visibleStepsWithConversionMetrics } = useValues(funnelDataLogic(insightProps))
     const { canOpenPersonModal } = useValues(funnelPersonsModalLogic(insightProps))
     const showPersonsModal = canOpenPersonModal && showPersonsModalProp
-    const chartId = useRef(`funnel-vertical-${Math.random().toString(36).substring(2, 11)}`)
-    const vizRef = useFunnelTooltip(showPersonsModal, chartId.current)
+    const vizRef = useFunnelTooltip(showPersonsModal)
 
     const { height: availableHeight } = useResizeObserver({ ref: vizRef })
     const [scrollbarHeightPx, setScrollbarHeightPx] = useState(0)
