+++ frontend/src/scenes/insights/views/WorldMap/WorldMap.tsx
@@ -9,6 +9,7 @@ import { COUNTRY_CODE_TO_LONG_NAME, countryCodeToFlag } from 'lib/utils/geograph
 import { InsightTooltip } from 'scenes/insights/InsightTooltip/InsightTooltip'
 import { formatAggregationAxisValue } from 'scenes/insights/aggregationAxisFormat'
 import { insightLogic } from 'scenes/insights/insightLogic'
+import { useInsightTooltip } from 'scenes/insights/useInsightTooltip'
 import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
 
 import { groupsModel } from '~/models/groupsModel'
@@ -17,7 +18,6 @@ import { QueryContext } from '~/queries/types'
 import { ChartParams, TrendResult } from '~/types'
 
 import { SeriesDatum } from '../../InsightTooltip/insightTooltipUtils'
-import { ensureTooltip } from '../LineGraph/LineGraph'
 import { countryVectors } from './countryVectors'
 import { worldMapLogic } from './worldMapLogic'
 
@@ -26,7 +26,7 @@ const SATURATION_FLOOR = 0.2
 /** The tooltip is offset by a few pixels from the cursor to give it some breathing room. */
 const WORLD_MAP_TOOLTIP_OFFSET_PX = 8
 
-function useWorldMapTooltip(showPersonsModal: boolean, chartId: string): React.RefObject<SVGSVGElement> {
+function useWorldMapTooltip(showPersonsModal: boolean): React.RefObject<SVGSVGElement> {
     const { insightProps } = useValues(insightLogic)
     const { series, trendsFilter, breakdownFilter, isTooltipShown, currentTooltip, tooltipCoordinates } = useValues(
         worldMapLogic(insightProps)
@@ -36,7 +36,8 @@ function useWorldMapTooltip(showPersonsModal: boolean, chartId: string): React.R
     const svgRef = useRef<SVGSVGElement>(null)
 
     const svgRect = svgRef.current?.getBoundingClientRect()
-    const [tooltipRoot, tooltipEl] = ensureTooltip(chartId)
+    const { getTooltip } = useInsightTooltip()
+    const [tooltipRoot, tooltipEl] = getTooltip()
 
     useEffect(() => {
         tooltipEl.style.opacity = isTooltipShown ? '1' : '0'
@@ -211,8 +212,7 @@ export function WorldMap({ showPersonsModal = true, context }: ChartParams): JSX
     const { countryCodeToSeries, maxAggregatedValue, querySource, theme } = useValues(worldMapLogic(insightProps))
     const { showTooltip, hideTooltip, updateTooltipCoordinates } = useActions(worldMapLogic(insightProps))
 
-    const chartId = useRef(`worldmap-${Math.random().toString(36).substring(2, 11)}`)
-    const svgRef = useWorldMapTooltip(showPersonsModal, chartId.current)
+    const svgRef = useWorldMapTooltip(showPersonsModal)
 
     const backgroundColor = theme?.['preset-1'] || '#000000' // Default to black if no color found
